+++ strings.py
